add disconnected sealer/observer in nodelist
