+++ src/main/java/com/webank/webase/front/base/response/BaseResponse.java
@@ -21,7 +21,7 @@ import lombok.Data;
 
 /**
  * BaseResponse.
- * 
+ * includi
  */
 @Data
 public class BaseResponse {
