+++ src/main/java/com/webank/webase/front/precompiledapi/permission/PermissionHandle.java
@@ -25,7 +25,6 @@ public class PermissionHandle {
     // defaultValue = "1"
     @NotNull(message = ConstantCode.PARAM_FAIL_GROUPID_IS_EMPTY)
     private int groupId;
-    // @NotNull(message = ConstantCode.PARAM_FAIL_PERMISSION_TYPE_IS_EMPTY)
     private String permissionType;
     @NotNull(message = ConstantCode.PARAM_FAIL_FROM_IS_EMPTY)
     private String fromAddress;
