+++ src/main/java/com/webank/webase/front/precompiledapi/permission/PermissionManageController.java
@@ -19,8 +19,8 @@ import com.alibaba.fastjson.JSON;
 import com.fasterxml.jackson.core.JsonParseException;
 import com.webank.webase.front.base.code.ConstantCode;
 import com.webank.webase.front.base.controller.BaseController;
-import com.webank.webase.front.base.entity.BasePageResponse;
-import com.webank.webase.front.base.entity.BaseResponse;
+import com.webank.webase.front.base.response.BasePageResponse;
+import com.webank.webase.front.base.response.BaseResponse;
 import com.webank.webase.front.base.exception.FrontException;
 import com.webank.webase.front.util.Address;
 import com.webank.webase.front.util.AddressUtils;
