+++ src/main/java/com/webank/webase/front/precompiledapi/permission/PermissionState.java
@@ -22,7 +22,7 @@ import javax.validation.constraints.NotNull;
 
 @Data
 public class PermissionState {
-    // 0 means revoke, 1 means grant
+    // 0 means revoked, 1 means granted
     @NotNull(message = ConstantCode.PARAM_FAIL_PERMISSION_STATE_ALL_CONNOT_BE_EMPTY)
     private int deployAndCreate;
     @NotNull(message = ConstantCode.PARAM_FAIL_PERMISSION_STATE_ALL_CONNOT_BE_EMPTY)
