+++ src/main/java/com/webank/webase/front/precompiledapi/sysconf/PrecompiledSysConfigController.java
@@ -15,8 +15,8 @@
  */
 package com.webank.webase.front.precompiledapi.sysconf;
 
-import com.webank.webase.front.base.entity.BasePageResponse;
-import com.webank.webase.front.base.entity.BaseResponse;
+import com.webank.webase.front.base.response.BasePageResponse;
+import com.webank.webase.front.base.response.BaseResponse;
 import com.webank.webase.front.base.code.ConstantCode;
 import com.webank.webase.front.util.PrecompiledUtils;
 import com.webank.webase.front.util.pageutils.List2Page;
